Simplify gridfeed logic
